refactor(lybic): split Stats class into separate module
- Move Stats class from lybic.py to new stats.py module
- Update import in __init__.py to reflect new module structure
- Improve code organization and maintainability
